mask_rcnn/requirements.txt
absl-py==0.15.0
alabaster==0.7.13
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astunparse==1.6.3
async-lru==1.0.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cycler==0.12.1
cython==0.29.24
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
docutils==0.20.1
exceptiongroup==1.2.2
fastjsonschema==2.21.1
flatbuffers==1.12
fqdn==1.5.1
gast==0.3.3
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h11==0.14.0
h5py==2.10.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.35.1
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipyparallel==9.0.0
ipython==7.34.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.4
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
keras-preprocessing==1.1.2
kiwisolver==1.4.7
markdown==3.7
markupsafe==2.1.5
matplotlib==3.3.4
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nose==1.3.7
notebook==7.3.1
notebook-shim==0.2.4
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.5.4.60
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==8.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.48
protobuf==3.20.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pywavelets==1.4.1
pywin32==308
pyyaml==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
qtpy==2.4.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9
scikit-image==0.17.2
scipy==1.5.4
send2trash==1.8.3
shapely==2.0.6
six==1.15.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.18.1
testpath==0.6.0
tifffile==2023.7.10
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==3.7.4.3
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.12.1
zipp==3.20.2